backend/requirements.txt
# Core FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
pymongo==4.9.0
motor==3.6.0

# Data validation and processing
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# Testing framework
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0

# Additional testing utilities
pytest-xdist==3.6.0
pytest-html==4.1.1

# Code quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Security
cryptography==44.0.0
